+++ packages/core/config/webpack/plugins.js
@@ -109,7 +109,6 @@ module.exports = function getPlugins(context, argv) {
           filename: '[name].css',
           chunkFilename: '[id].css',
         }),
-
       ];
 
     default: