+++ packages/styled-components/CHANGELOG.md
@@ -2,6 +2,276 @@
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+## [5.8.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.8.0...v5.8.1) (2021-03-23)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+## [5.8.1-alpha.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.8.0...v5.8.1-alpha.0) (2021-03-23)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+# [5.8.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.8.0-alpha.2...v5.8.0) (2021-03-23)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.8.0-alpha.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.8.0-alpha.0...v5.8.0-alpha.1) (2021-03-23)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.8.0-alpha.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.7.0...v5.8.0-alpha.0) (2021-03-16)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.7.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.7.0-beta.0...v5.7.0) (2021-03-16)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.7.0-beta.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.7.0-alpha.4...v5.7.0-beta.0) (2021-03-12)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.7.0-alpha.4](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.7.0-alpha.3...v5.7.0-alpha.4) (2021-03-12)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.7.0-alpha.3](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.6.1...v5.7.0-alpha.3) (2021-03-02)
+
+
+
+# [5.7.0-alpha.2](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.7.0-alpha.1...v5.7.0-alpha.2) (2021-02-25)
+
+
+### Bug Fixes
+
+* **loader:** make loader configurable and themable via bodyWrapper ([a534d16](https://github.com/alleyinteractive/irving/packages/styled-components/commit/a534d16ca7ecb6b61fa6a7e289b27a0c2f92c7f0))
+* **loader:** remove test case for loading component, remove fullscreen option/prop ([c1e4d6f](https://github.com/alleyinteractive/irving/packages/styled-components/commit/c1e4d6f1817589240bd3f1d4429a8c957140a0a3))
+
+
+
+# [5.7.0-alpha.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.7.0-alpha.0...v5.7.0-alpha.1) (2021-02-22)
+
+
+
+# [5.7.0-alpha.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.6.0...v5.7.0-alpha.0) (2021-02-22)
+
+
+
+
+
+# [5.7.0-alpha.2](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.7.0-alpha.1...v5.7.0-alpha.2) (2021-02-25)
+
+
+### Bug Fixes
+
+* **loader:** make loader configurable and themable via bodyWrapper ([a534d16](https://github.com/alleyinteractive/irving/packages/styled-components/commit/a534d16ca7ecb6b61fa6a7e289b27a0c2f92c7f0))
+* **loader:** remove test case for loading component, remove fullscreen option/prop ([c1e4d6f](https://github.com/alleyinteractive/irving/packages/styled-components/commit/c1e4d6f1817589240bd3f1d4429a8c957140a0a3))
+
+
+
+
+
+# [5.7.0-alpha.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.7.0-alpha.0...v5.7.0-alpha.1) (2021-02-22)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.7.0-alpha.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.6.0...v5.7.0-alpha.0) (2021-02-22)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+## [5.6.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.6.1-beta.0...v5.6.1) (2021-03-02)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+## [5.6.1-beta.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.6.0...v5.6.1-beta.0) (2021-03-01)
+
+
+### Bug Fixes
+
+* **loader:** remove test case for loading component, remove fullscree… … ([92eb9b0](https://github.com/alleyinteractive/irving/packages/styled-components/commit/92eb9b0607a1c14963691bd3ee9dc2a664c21740))
+
+
+
+
+
+# [5.6.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.6.0-beta.0...v5.6.0) (2021-02-22)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.6.0-beta.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.6.0-alpha.1...v5.6.0-beta.0) (2021-02-19)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.6.0-alpha.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.6.0-alpha.0...v5.6.0-alpha.1) (2021-02-19)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.6.0-alpha.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.11...v5.6.0-alpha.0) (2021-02-19)
+
+
+
+# [5.5.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-beta.1...v5.5.0) (2021-02-17)
+
+
+
+# [5.5.0-beta.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.10...v5.5.0-beta.1) (2021-02-17)
+
+
+
+# [5.5.0-beta.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.9...v5.5.0-beta.0) (2021-02-16)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.5.0-alpha.11](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.10...v5.5.0-alpha.11) (2021-02-19)
+
+
+### Bug Fixes
+
+* **providers:** make it so providers using route key always update on location change ([fbf1dad](https://github.com/alleyinteractive/irving/packages/styled-components/commit/fbf1dad63d3339715a7cc353945e1ec4ae50a9cd))
+* **providers:** use switch statement for providersReducer ([c0fc009](https://github.com/alleyinteractive/irving/packages/styled-components/commit/c0fc0093540242636fbb85576fe5865ad4972c70))
+
+
+### Reverts
+
+* **head:** revert head component to the way it was pre-provider change ([#462](https://github.com/alleyinteractive/irving/packages/styled-components/issues/462)) ([86b5fb1](https://github.com/alleyinteractive/irving/packages/styled-components/commit/86b5fb1885736ba75adbbf9dba6f2e8a768f5361))
+
+
+
+
+
+# [5.5.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-beta.1...v5.5.0) (2021-02-17)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.5.0-beta.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.10...v5.5.0-beta.1) (2021-02-17)
+
+
+
+# [5.5.0-beta.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.9...v5.5.0-beta.0) (2021-02-16)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.5.0-beta.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.9...v5.5.0-beta.0) (2021-02-16)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+# [5.5.0-alpha.10](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.9...v5.5.0-alpha.10) (2021-02-17)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.5.0-alpha.9](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.8...v5.5.0-alpha.9) (2021-02-16)
+
+
+### Bug Fixes
+
+* **providers:** auto-convert provider apiChildren to components ([c5950c9](https://github.com/alleyinteractive/irving/packages/styled-components/commit/c5950c965375717485564ba2717bf61df7b0e0a3))
+
+
+### Features
+
+* **provider:** change head component into provider, allow providers to have children ([085338c](https://github.com/alleyinteractive/irving/packages/styled-components/commit/085338c681ef402f93619df4066c095550bcf806))
+
+
+
+
+
+# [5.5.0-alpha.8](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.7...v5.5.0-alpha.8) (2021-02-16)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.5.0-alpha.7](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.6...v5.5.0-alpha.7) (2021-02-10)
+
+
+
+## [5.4.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.4.1-rc.3...v5.4.1) (2021-02-05)
+
+
+
+## [5.4.1-rc.3](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.5...v5.4.1-rc.3) (2021-02-04)
+
+
+
+## [5.4.1-rc.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.4...v5.4.1-rc.1) (2021-02-02)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [5.5.0-alpha.6](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.5...v5.5.0-alpha.6) (2021-02-10)
+
+
+### Features
+
+* **text:** add `nowrap` prop to text component to prevent rendering outer div ([23a64c4](https://github.com/alleyinteractive/irving/packages/styled-components/commit/23a64c40baff0af38c6271c1f8de01b50c8d48e5))
+
+
+
+
 
 # [5.5.0-alpha.5](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.4...v5.5.0-alpha.5) (2021-02-04)
 
@@ -30,6 +300,9 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 # [5.5.0-alpha.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.5.0-alpha.0...v5.5.0-alpha.1) (2021-01-28)
 
 **Note:** Version bump only for package @irvingjs/styled-components
+## [5.4.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.4.1-rc.3...v5.4.1) (2021-02-05)
+
+**Note:** Version bump only for package @irvingjs/styled-components
 
 
 
@@ -38,12 +311,18 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 # [5.5.0-alpha.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.4.1-alpha.0...v5.5.0-alpha.0) (2021-01-27)
 
 **Note:** Version bump only for package @irvingjs/styled-components
+## [5.4.1-rc.3](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.4.1-rc.2...v5.4.1-rc.3) (2021-02-04)
+
+**Note:** Version bump only for package @irvingjs/styled-components
 
 
 
 
 
 ## [5.4.1-alpha.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.4.0...v5.4.1-alpha.0) (2021-01-27)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+## [5.4.1-rc.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v5.4.1-rc.0...v5.4.1-rc.1) (2021-02-02)
 
 **Note:** Version bump only for package @irvingjs/styled-components